Add an environment bootstrapping step to readme.md
In order to line up with the `Environment Setup` wiki page, we need to
ensure that users have run the bootstrapping script and are able to use
the subcommands under `jet`. To do this, the `How to pull` section
needed to be moved before the `How to get the Docker image` section. In
a later commit, the rest of the commands in `readme.md` will be changed
to `jet` subcommands.