Move more involved readme.md sections to the wiki
Both the `Using the shell inside the Docker container` and the `Using
our Build-System` sections have been moved to the wiki under pages with
those or slightly altered names. Since the `readme.md` file is meant to
be a quick-start entry point into the project, it should be short and
simple.