Add "--qemu-arm" flag to "jet build" and "jet run"
Building on the ODROID appears to take longer than the average human
life span, so this flag enables using QMEU User Emulation to run the
build (and potentially the whole stack) in the `arm` container on an
`x86_64` machine.

If this flag is used, `get_image_name` will verify that the
`qemu-user-static` binary is present on the system and that the user is
indeed on an `x86_64` platform. If these checks pass, the
`hoverjet/jet-arm` image will be fetched and the
`/usr/bin/qemu-user-static` binary will be added as a bind mount to the
`docker run` commands used in the `jet_build.sh` and `jet_run.sh`
scripts. From there on, everything should be seamless.

Preliminary testing has just consisted of building the full project,
which seems to take very close to 96 minutes regardless of the number of
jobs used (valused of `3`, `4`, and `7` were tested). This implies that
the build does not parallelize well and that throwing more cores at the
problem will not help much, but it should still be tested. The machine
that ran this testing had `2` cores and `4` threads.

Running the software stack both within a container running in this mode
and on the jet itself (the build repository would be copied over) should
also be tested prior to landing.